requirements.txt
asgiref==3.2.7
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
Markdown==3.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
